doc: drop mention of hypothetical "IPFN".
It's no good to introduce a term just to spend a halfhearted paragraph
partially describing it just to say that it's out of scope.

The schema page also nowadays already has some mentions of how schema
are designed to avoid turing completeness and are not intended to
encompass dependent typing systems, so this descoping is already noted.

Signed-off-by: Eric Myhre <hash@exultant.us>