An earlier draft of an error type.
It's high time we got started on well-typed errors; and this type name
is already hinted at by the codegen outputs, so let's make it real.

This file has been hanging around on my disk uncommitted for a fair
while, and some of the comments about review-needed are stale, but I'm
keepin' em for historical interest.  I think I have fixes planned that
will straight out all of them.