Confess the config file for localhost CI.
This works via the 'reach' tool (and more specifically,
'reach ci' subcommand) from the Timeless Stack.

(The catalog files for the compiler and base image should perhaps be
vendored as well for maximum reliability, but it also works to create
a '.timeless' dir somewhere either in or above this working directory,
and clone the https://github.com/polydawn/catalog repo to the
'.timeless/catalog' path.)

Note that while this works for "localhost CI" usecases, we'll also be
using some other tooling for getting the familiar integrations with
Github and online dashboards, etc.  These can coexist in parallel.